+++ pkg/providers/property/testdata/TestResourceEdgeHostname/invalid_edgehostname_domainprefix.tf
@@ -1,0 +1,11 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+resource "akamai_edge_hostname" "edgehostname" {
+  contract_id   = "ctr_2"
+  group_id      = "grp_2"
+  product_id    = "prd_2"
+  edge_hostname = "testABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz01234567.edgesuite.net"
+  ip_behavior   = "IPV6_PERFORMANCE"
+}