+++ CHANGELOG.md
@@ -18,6 +18,9 @@
 * PAPI
   * Added populating `product_id` during import of `akamai_edge_hostname` if available and not provided in import key.
   * Added the `akamai_cp_codes` datasource. ([I#604](https://github.com/akamai/terraform-provider-akamai/issues/604)).
+  * Added name format validations and a size limit for the Domain Prefix.
+
+